+++ src/dataprotection/azext_dataprotection/manual/Manifests/AzureKubernetesService.py
@@ -52,7 +52,7 @@ manifest = '''
   ],
   "policySettings": {
     "supportedRetentionTags": [ "Daily", "Weekly" ],
-    "supportedDatastoreTypes": [ "OperationalStore" ],
+    "supportedDatastoreTypes": [ "OperationalStore", "VaultStore" ],
     "disableAddRetentionRule": false,
     "disableCustomRetentionTag": true,
     "backupScheduleSupported": true,
@@ -71,6 +71,22 @@ manifest = '''
               ]
             },
             "taggingCriteria": [
+              {
+                "tagInfo": {
+                  "tagName": "Daily",
+                  "id": "Daily_"
+                },
+                "taggingPriority": 25,
+                "isDefault": false,
+                "criteria": [
+                  {
+                    "absoluteCriteria": [
+                      "FirstOfDay"
+                    ],
+                    "objectType": "ScheduleBasedBackupCriteria"
+                  }
+                ]
+              },
               {
                 "tagInfo": {
                   "tagName": "Default",
@@ -104,6 +120,45 @@ manifest = '''
           ],
           "isDefault": true,
           "name": "Default",
+          "objectType": "AzureRetentionRule"
+        },
+        {
+          "lifecycles": [
+            {
+              "deleteAfter": {
+                "objectType": "AbsoluteDeleteOption",
+                "duration": "P7D"
+              },
+              "targetDataStoreCopySettings": [
+                {
+                  "dataStore": {
+                    "dataStoreType": "VaultStore",
+                    "objectType": "DataStoreInfoBase"
+                  },
+                  "copyAfter": {
+                    "objectType": "ImmediateCopyOption"
+                  }
+                }
+              ],
+              "sourceDataStore": {
+                  "dataStoreType": "OperationalStore",
+                  "objectType": "DataStoreInfoBase"
+              }
+            },
+            {
+              "deleteAfter": {
+                "objectType": "AbsoluteDeleteOption",
+                "duration": "P84D"
+              },
+              "targetDataStoreCopySettings": [],
+              "sourceDataStore": {
+                "dataStoreType": "VaultStore",
+                "objectType": "DataStoreInfoBase"
+              }
+            }
+          ],
+          "isDefault": false,
+          "name": "Daily",
           "objectType": "AzureRetentionRule"
         }
       ],